undo top level tab nav shuffle
